Split the Message and Group Controllers Into Sub Modules for Maintainability
