+++ server/controllers/group/index.js
@@ -1,0 +1,11 @@
+import ModuleOneGroupController from './module-one.group';
+import ModuleTwoGroupController from './module-two.group';
+import ModuleThreeGroupController from './module-three.group';
+
+const groupController = {
+  moduleOne: ModuleOneGroupController,
+  moduleTwo: ModuleTwoGroupController,
+  moduleThree: ModuleThreeGroupController,
+};
+
+export default groupController;
