+++ server/controllers/message/index.js
@@ -1,0 +1,11 @@
+import ModuleOneMessageController from './module-one.message';
+import ModuleTwoMessageController from './module-two.message';
+import ModuleThreeMessageController from './module-three.message';
+
+const messageController = {
+  moduleOne: ModuleOneMessageController,
+  moduleTwo: ModuleTwoMessageController,
+  moduleThree: ModuleThreeMessageController,
+};
+
+export default messageController;
