+++ server/routes/message.js
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 
 import email from '../middleware/validate';
-import MessageController from '../controllers/message';
+import messageController from '../controllers/message';
 import authentication from '../middleware/authenticate';
 
 const router = Router();
@@ -10,79 +10,79 @@ router.post(
   '/messages',
   authentication.verifyUser,
   email.validateEmail,
-  MessageController.sendEmail
+  messageController.moduleOne.sendEmail
 );
 
 router.get(
   '/messages',
   authentication.verifyUser,
-  MessageController.retrieveReceivedEmails
+  messageController.moduleOne.retrieveReceivedEmails
 );
 
 router.get(
   '/messages/sent',
   authentication.verifyUser,
-  MessageController.retrieveSentEmails
+  messageController.moduleOne.retrieveSentEmails
 );
 
 router.get(
   '/messages/read',
   authentication.verifyUser,
-  MessageController.retrieveReadEmails
+  messageController.moduleThree.retrieveReadEmails
 );
 
 router.get(
   '/messages/unread',
   authentication.verifyUser,
-  MessageController.retrieveUnReadEmails
+  messageController.moduleThree.retrieveUnReadEmails
 );
 
 router.get(
   '/messages/draft',
   authentication.verifyUser,
-  MessageController.retrieveDraftEmails
+  messageController.moduleThree.retrieveDraftEmails
 );
 
 router.get(
   '/messages/:id',
   authentication.verifyUser,
   email.validateEmailId,
-  MessageController.retrieveSpecificReceivedEmail
+  messageController.moduleOne.retrieveSpecificReceivedEmail
 );
 
 router.delete(
   '/messages/:id',
   authentication.verifyUser,
   email.validateEmailId,
-  MessageController.deleteSpecificReceivedEmail
+  messageController.moduleTwo.deleteSpecificReceivedEmail
 );
 
 router.get(
   '/messages/sent/:id',
   authentication.verifyUser,
   email.validateEmailId,
-  MessageController.retrieveSpecificSentEmail
+  messageController.moduleThree.retrieveSpecificSentEmail
 );
 
 router.get(
   '/messages/draft/:id',
   authentication.verifyUser,
   email.validateEmailId,
-  MessageController.retrieveSpecificDraftEmail
+  messageController.moduleTwo.retrieveSpecificDraftEmail
 );
 
 router.delete(
   '/messages/sent/:id',
   authentication.verifyUser,
   email.validateEmailId,
-  MessageController.deleteSpecificSentEmail
+  messageController.moduleTwo.deleteSpecificSentEmail
 );
 
 router.delete(
   '/messages/draft/:id',
   authentication.verifyUser,
   email.validateEmailId,
-  MessageController.deleteSpecificDraftEmail
+  messageController.moduleTwo.deleteSpecificDraftEmail
 );
 
 export default router;
